server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
fastapi==0.116.1
h11==0.16.0
HandTrackingModule==0.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.74
langchain-openai==0.3.28
langsmith==0.4.13
motor==3.7.1
numpy==2.2.6
openai==1.99.3
opencv-python==4.12.0.88
orjson==3.11.1
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.10.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
zstandard==0.23.0
